API Reference

Models

Power spectral densities

Tonari.DoubleBendingPowerLaw — Type

 DoubleBendingPowerLaw(α₁, f₁, α₂, f₂, α₃)

Double bending power law model for the power spectral density

α₁: the first power law index

f₁: the first bend frequency

α₂: the second power law index

f₂: the second bend frequency

α₃: the third power law index

\[\mathcal{P}(f) = A\frac{(f/f₁)^{-α₁}}{1 + (f / f₁)^{α₂ - α₁}}\frac{1}{1 + (f / f₂)^{α₃ - α₂}}\]

Tonari.Lorentzian — Type

Lorentzian(A, γ, f₀)

Lorentzian model for the power spectral density

A: the amplitude

γ: the width of the peak

f₀: the central frequency

\[\mathcal{P}(f) = \frac{A}{4\pi^2 (f - f₀)^2 + γ^2}\]

Tonari.SingleBendingPowerLaw — Type

 SingleBendingPowerLaw(A, α₁, f₁, α₂)

Single bending power law model for the power spectral density

A: the amplitude

α₁: the first power law index

f₁: the first bend frequency

α₂: the second power law index

\[\mathcal{P}(f) = A \frac{(f/f₁)^{-α₁}}{1 + (f / f₁)^{α₂ - α₁}}\]

Simulations

Tonari.Simulation — Type

Simulation(model::Model, T::Real, Δt::Real, S_high::Real, S_low::Real, t::AbstractVector{Real})

A struct that contains the information of a simulation of a stochastic process.

Fields

model::Model: The power spectral density of the process.

T::Real: The duration of the simulated time series. Note that the time stamps are from 0 to T, so the duration is T.

Δt::Real: The sampling period, or minimum time difference between samples.

S_high::Real: The factor by which the maximum frequency is multiplied for the simulation.

S_low::Real: The factor by which the minimum frequency is divided for the simulation.

t::AbstractVector{Real}: The time vector, in this case, the time at which the process is sampled. It is assumed that the time vector is sorted.

Constructors

Simulation(model::Model, T::Real, Δt::Real, S_high::Real, S_low::Real, t::AbstractVector{Real}): Constructs a simulation with regular sampling.

Simulation(model::Model, T::Real, Δt::Real): Constructs a simulation with regular sampling, and sets S_high and S_low to 10.0.

Simulation(model::Model, T::Real, Δt::Real, S_high::Real, S_low::Real): Constructs a simulation with the given parameters.

Simulation(model::Model, t::AbstractVector{Real}, S_high::Real, S_low::Real): Constructs a simulation with sampling pattern given by t.

Simulation(model::Model, t::AbstractVector{Real}): Constructs a simulation with sampling pattern given by t, and sets S_high and S_low to 10.0.

StatsBase.sample — Function

sample(rng, sim, n=1, input_mean=0; σₓ = nothing, split_long=false, randomise_values=true, Fvar=nothing, alt=false, poisson=false, exponentiate=false, error_size=0.02)

Generate a time series with a given power spectral density (PSD) using the (Timmer and König, 1995) method.

Arguments

rng::MersenneTwister: Random number generator.

sim::Simulation: The simulation

n::Int: Number of time series to generate. Default is 1.

input_mean::Real: The mean of the time series. Default is 0.

σₓ::Array{Float64, 1}: The errorbars of the time series. Default is nothing.

split_long::Bool: If true, the time series is split into shorter time series given by sim.S_low-1. Default is true.

randomise_values::Bool: If true, the values of the time series are randomised. Default is true.

Fvar::Real: The variance of the time series. Default is nothing.

alt::Bool: If true, uses the alternative Timmer & Koenig method. Default is false.

poisson::Bool: If true, Poisson noise is added to the time series. Default is false.

exponentiate::Bool: If true, the time series is exponentiated. Default is false.

error_size::Real: The size of the error. Default is 0.05.

Tonari.draw_errorbars — Function

draw_errorbars(rng, x, Δt, poisson=false, error_size=0.05)

Draw errorbars for the time series.

Arguments

rng::MersenneTwister: Random number generator.

x::Array{Float64, 1}: The values of the time series.

Δt::Real: The sampling period.

poisson::Bool: If true, Poisson noise is added to the time series. Default is false.

error_size::Real: The size of the error. Default is 0.05.

Tonari.findnearest — Method

findnearest(a, b)

Find the nearest value in b to each value in a. This assumes that a and b are sorted.

Tonari.randomise_fluxes — Function

randomise_fluxes(rng, xₛ, Δt, input_mean = 0.0; σ = nothing, Fvar = nothing, poisson = false, exponentiate = false, error_size = 0.05)

Randomise the fluxes of the time series.

Arguments

rng::MersenneTwister: Random number generator.

xₛ::Array{Float64, 1}: The values of the time series.

Δt::Real: The sampling period.

input_mean::Real: The mean of the time series. Default is 0.

σ::Array{Float64, 1}: The errorbars of the time series. Default is nothing.

Fvar::Real: The variance of the time series. Default is nothing.

poisson::Bool: If true, Poisson noise is added to the time series. Default is false.

exponentiate::Bool: If true, the time series is exponentiated. Default is false.

Tonari.sample_split_timeseries — Method

sample_split_timeseries(x, t, t_desired, n_sim, n, n_slices, split_long)

Split the time series into shorter time series and sample at the desired time stamps.

Arguments

x::Array{Float64, 1} or Vector{Matrix{Float64}}: The values of the time series. If there are multiple bands, the time series is a vector of matrices.

t::Array{Float64, 1}: The time indexes of the time series.

t_desired::Array{Float64, 1}: The desired time stamps.

n_sim::Int: The number of simulations.

n::Int: The number of time series to generate.

n_slices::Int: The number of slices to break the time series into.

split_long::Bool: If true, the time series is split into shorter time series. Default is true.

Returns

A tuple of two arrays: the first containing the time indexes of the shorter time series, and the second containing the values of the shorter time series.

Tonari.sample_timeseries — Method

sample_timeseries(t, y, M)

Extract a random subset of points from the time series.

Tonari.split_longtimeseries — Method

split_longtimeseries(t, ts, n_slices, t_end)

Split a long time series into shorter time series. Break the time series into n_slices shorter time series. The short time series are of equal length.

Arguments

t: The time indexes of the long time series.

ts: The values of the long time series.

n_slices: The number of slices to break the time series into.

t_end

Returns

A tuple of two lists: the first containing the time indexes of the shorter time series, and the second containing the values of the shorter time series.

Tonari.timmer_koenig — Method

timmer_koenig(psd, rng::random.AbstractRNG, alternative=false)

Generate a time series with a given power spectral density (PSD) using the (Timmer and König, 1995) method

Given N values of the power spectral density (PSD) 𝓟

Draw 2N values from a standard normal distribution.

The amplitude of the randomised periodogram is given by A = N + i M

The randomised periodogram is given by 𝓟_rand = √(𝓟 / 2) * A

The first value of the randomised periodogram inserted and set to 0

The time series is obtained by taking the inverse Fourier transform of the randomised periodogram, with the function FFTW.irfft. We use the length of the time series as 2*(N+1)-1.

The alternative parametrisation is given by

Given N values of the power spectral density (PSD) 𝓟

Draw N-1 values from a χ²(2) distribution and 1 value from χ²₁(1), this is A the amplitude of the randomised periodogram

Draw N values from a uniform distribution between 0 and 1, and set the last value to 0, this is θ the phase of the randomised periodogram

The randomised periodogram is given by 𝓟_rand = √(𝓟 / 2 * A) * exp(2πiθ)

The first value of the randomised periodogram inserted and set to 0

The time series is obtained by taking the inverse Fourier transform of the randomised periodogram, with the function FFTW.irfft. We use the length of the time series as 2*(N+1)-1.

Arguments

𝓟::Array{Float64, 1}: PSD associated with the process

rng::Random.AbstractRNG: Random number generator.

α::Real: Multiplicative factor for the randomised periodogram. Default is 1.0. This can be a complex vector to add a phase to the time series.

Returns

x::Array{Float64, 1}: Time series with the given PSD.

Periodogram

Tonari.coherence — Function

coherence(f, C̄, P̄₁, P̄₂, n_segments, y₁_err = nothing, y₂_err = nothing)

Compute the coherence between two time series defined in (Vaughan and Nowak, 1997) using the cross-periodogram C̄, the mean power spectrum of the first time series P̄₁, and the mean power spectrum of the second time series P̄₂. The coherence is defined as:

\[\gamma^2 (f) = \frac{|\overline{C}(f)|^2}{\overline{P}_1(f) \overline{P}_2(f)}\]

with the error in the coherence given by:

\[\sigma_{\gamma^2} = \sqrt{2} \sqrt{\gamma^2} \frac{1 - \gamma^2}{\sqrt{M}}\]

where M is the number of segments over which the coherence is computed.

The phase lag is defined as:

\[\Delta \phi (f) = \arg \overline{C}(f)\]

with the error in the phase lag given by:

\[\sigma_{\Delta \phi} = \frac{1 - \gamma^2}{\sqrt{2 \gamma^2 M}}\]

The time lag is defined as:

\[\Delta \tau (f) = \frac{\Delta \phi}{2 \pi f}\]

with the error in the time lag given by:

\[\sigma_{\Delta \tau} = \frac{\sigma_{\Delta \phi}}{2 \pi f}\]

If the errors in the time series are provided, the corrected coherence is computed using the method described in (Vaughan and Nowak, 1997).

Arguments

f::Array{Float64, 1}: Frequency array.

C̄::Array{Complex{Float64}, 1}: Cross-periodogram.

P̄₁::Array{Float64, 1}: Mean power spectrum of the first time series.

P̄₂::Array{Float64, 1}: Mean power spectrum of the second time series.

n_segments::Int: Number of segments over which the coherence is computed.

y₁_err::Array{Float64, 1}: Error in the first time series.

y₂_err::Array{Float64, 1}: Error in the second time series.

if the errors in the time series are not provided, the function returns the coherence, phase lag, and time lag. Otherwise, it returns the corrected coherence, phase lag, and time lag.

Returns

f::Array{Float64, 1}: Frequency array.

γ²::Array{Float64, 1}: Coherence.

Δφ::Array{Float64, 1}: Phase lag.

γ²_err::Array{Float64, 1}: Error in the coherence.

Δφ_err::Array{Float64, 1}: Error in the phase lag.

Δτ::Array{Float64, 1}: Time lag.

Δτ_err::Array{Float64, 1}: Error in the time lag.

P̄₁::Array{Float64, 1}: Mean power spectrum of the first time series.

Tonari.cross_periodogram — Function

cross_periodogram(t, y₁, y₂, y₁_err = nothing, y₂_err = nothing; compute_coherence = true, apply_end_matching = false, subtract_mean = true)

Compute the cross-periodogram between two time series y₁ and y₂ with time stamps t using the fast Fourier transform (FFT).

Arguments

t::Array{Float64, 1}: Time stamps.

y₁::Array{Float64, 1}: First time series data. Or a matrix of multiple time series.

y₂::Array{Float64, 1}: Second time series data. Or a matrix of multiple time series.

y₁_err::Array{Float64, 1}: Error in the first time series. Or a matrix of errors for multiple time series.

y₂_err::Array{Float64, 1}: Error in the second time series Or a matrix of errors for multiple time series.

compute_coherence::Bool: Compute the coherence between the two time series. Default is true.

apply_end_matching::Bool: Apply end-matching to the data. Default is false.

subtract_mean::Bool: Subtract the mean from the data. Default is true.

If compute_coherence is true, the function returns the coherence between the two time series. Otherwise, it returns the cross-periodogram and the mean power spectra of the two time series. see coherence function for the optional returns.

Returns

f::Array{Float64, 1}: Frequency array.

C̄::Array{Complex{Float64}, 1}: Mean cross-periodogram.

P̄₁::Array{Float64, 1}: Mean power spectrum of the first time series.

P̄₂::Array{Float64, 1}: Mean power spectrum of the second time series.

C::Array{Complex{Float64}, 1}: Cross-periodogram.

Tonari.end_matching — Method

end_matching(y,t)

End-match the data y with a straight line and subtract it from the data.

Tonari.periodogram — Function

periodogram(t,y, normalisation = "default"; apply_end_matching = false, subtract_mean = true)

Compute the periodogram of the data y with time stamps t using the fast Fourier transform (FFT).

The periodogram is computed as the squared magnitude of the Fourier transform of the data. In practrice we use the real-valued fast Fourier transform (rfft) to compute the periodogram.

Arguments

t::Array{Float64, 1}: Time stamps.

y::Array{Float64, 1}: Time series data.

normalisation::Float64: Normalisation factor. Default is 2Δt / length(t).

apply_end_matching::Bool: Apply end-matching to the data. Default is false.

subtract_mean::Bool: Subtract the mean from the data. Default is true.

Utilities

Tonari.fill_gaps — Function

fill_gaps(rng::AbstractRNG, t::Vector{Float64}, x::Vector{Float64}, σₓ = nothing; randomise_values::Bool = true, poisson::Bool = true)

Fill gaps in a time series data with linear interpolation. If randomise_values = true, the interpolated values are drawn from a normal distribution with the mean and standard deviation of the data. If poisson = true, the interpolated values are drawn from a Poisson distribution with the mean of the data.

Arguments

rng::AbstractRNG: random number generator

t::Vector{Float64}: time array

x::Vector{Float64}: data array

σₓ::Vector{Float64}: standard deviation of the data array (optional)

randomise_values::Bool: whether to randomise the interpolated values

poisson::Bool: whether to use a Poisson distribution for the interpolated values, this assumes that the data is in counts/dt