API Reference

Time series

Tonari.IrregularTimeStamps — Type

IrregularTimeStamps

A structure to store the time stamps of a time series. It contains the time stamps, the time step, the duration of the time series and a boolean to indicate if the time stamps are irregular. If the time stamps are irregular, the time step is the minimum time step between two time stamps.

Fields

time::AbstractVector{T}: the time stamps

unit::Unitful.Unit: the unit of the time stamps (e.g. seconds, days, etc)

timezero::Td: the time of the first time stamp

Δt::T: the time step

duration::T: the duration of the time series

Tonari.RegularTimeStamps — Type

RegularTimeStamps

A structure to store the time stamps of a time series. It contains the time stamps, the time step, the duration of the time series and a boolean to indicate if the time stamps are irregular. If the time stamps are irregular, the time step is the minimum time step between two time stamps.

Fields

Δt::T: the time step

duration::T: the duration of the time series

unit::Unitful.Unit: the unit of the time stamps (e.g. seconds, days, etc)

timezero::Td: the time of the first time stamp

Tonari.TimeSeries — Type

TimeSeries

A structure to store a time series. It contains the time stamps, the data, and a dictionary to store metadata.

Fields

time::TimeStamps{Tt,Tz}: the time stamps

data::TimeSeriesData{Td, N}: the data

metadata::Dict{Symbol, Any}: a dictionary to store metadata

Entries in metadata

:name: a string to store the name of the time series

:description: a string to store a description of the time series

:unit: a string to store the unit(s) of the data

:unit_time: unit of the time stamps (default: seconds)

:timezero: a string to store the time of the first time stamp

:fake: a boolean to indicate if the time series is fake

:instrument: a string to store the instrument used to measure the time series

:irregular: a boolean to indicate if the time series has irregular time stamps

Tonari.TimeSeries — Method

TimeSeries(time::AbstractVector{T}, data::AbstractVector{Td}, metadata::Dict{Symbol,Any} = Dict{Symbol,Any}(), atol::T=0) where {T, Td}

A constructor for the TimeSeries structure.

This function lets you create a TimeSeries structure using the times, data, and optional metadata.

Arguments

time::AbstractVector{T}: the time stamps of the time series (length N)

data::AbstractVector{Td}: the data array (N x M) where N is the number of time stamps and M is the number of columns, M can be 1 for a single column, then data is a vector

metadata::Dict{Symbol,Any}: a dictionary to store metadata

atol::T=0: the absolute tolerance for the time stamps to check if they are regular

Metadata (optional but highly recommended)

:irregular::Bool : a boolean to indicate if the time stamps are irregular

:unit_time: the unit of the time stamps (default: seconds) using Unitful

:timezero: the time of the first time stamp (default: time[1])

:columns: the names of the columns (length M, default: y1, y2, ...), if M = 1, then the column is a Symbol

:unit: the units of the columns (using Unitful) (length M, default: dimensionless), if M = 1, then the unit is a Symbol, but if all columns have the same unit, then the unit is a single unit

:name: the name of the time series (optional)

:description: a description of the time series (optional)

:fake: a boolean to indicate if the time series is fake (optional)

:instrument: the instrument used to measure the time series (optional)

Returns

TimeSeries: a TimeSeries structure with the time stamps, data, and metadata

Tonari.TimeSeriesData — Type

TimeSeriesData

A type for the data in the time series.

Fields

Union{Vector{Symbol},Symbol}: the names of the columns, e.g. flux, error. Can be a vector of symbols or a single symbol if there is only one column.

data::Union{AbstractArray{T,2},Vector{T}}: the data array. Can be a matrix of data or a vector if there is only one column.

units::Union{Vector{Tu},Tu}: the units of the data. Can be a vector of units or a single unit if there is only one column or the same unit is used for all columns.

Tonari.fill_gaps — Function

fill_gaps(rng::AbstractRNG, t::Vector{Float64}, x::Vector{Float64}, σₓ = nothing; randomise_values::Bool = true, poisson::Bool = true)

Fill gaps in a time series data with linear interpolation. If randomise_values = true, the interpolated values are drawn from a normal distribution with the mean and standard deviation of the data. If poisson = true, the interpolated values are drawn from a Poisson distribution with the mean of the data.

Arguments

rng::AbstractRNG: random number generator

t::Vector{Float64}: time array

x::Vector{Float64}: data array

σₓ::Vector{Float64}: standard deviation of the data array (optional)

randomise_values::Bool: whether to randomise the interpolated values (default: true)

poisson::Bool: whether to use a Poisson distribution for the interpolated values, this assumes that the data is in counts/dt units (default: true)

Δt::Float64: the time step of the time series (optional)

Returns

t_filled::Vector{Float64}: the time array with the gaps filled

x_filled::Vector{Float64}: the data array with the gaps filled

σ_filled::Vector{Float64}: the standard deviation of the data array with the gaps filled (optional)

Tonari.time_series_sanity_checks — Method

time_series_sanity_checks(t₁::AbstractVector{T}, x₁::AbstractVector{T}, σ₁=nothing) where T

Perform sanity checks on the time series. The function checks the following:

The time and value vectors have the same length

The time and uncertainty vectors have the same length

The time series are sorted in ascending order

The time series do not contain infinities or NaNs

Arguments

t₁::AbstractVector{T}: time points of the time series

x₁::AbstractVector{T}: values of the time series

σ₁::AbstractVector{T}=nothing: uncertainty of the time series

Models

Power spectral densities

Tonari.DoubleBendingPowerLaw — Type

 DoubleBendingPowerLaw(A, α₁, f₁, α₂, f₂, α₃)

Double bending power law model for the power spectral density

A : the amplitude

α₁: the first power law index

f₁: the first bend frequency

α₂: the second power law index

f₂: the second bend frequency

α₃: the third power law index

\[\mathcal{P}(f) = A\frac{(f/f₁)^{-α₁}}{1 + (f / f₁)^{α₂ - α₁}}\frac{1}{1 + (f / f₂)^{α₃ - α₂}}\]

Tonari.Lorentzian — Type

Lorentzian(A, γ, f₀)

Lorentzian model for the power spectral density

A: the amplitude

γ: the width of the peak

f₀: the central frequency

\[\mathcal{P}(f) = \frac{A}{4\pi^2 (f - f₀)^2 + γ^2}\]

Tonari.PowerLaw — Type

 PowerLaw(α)

Power law model for the power spectral density

α: the power law index

\[\mathcal{P}(f) = A f^{-α}\]

Tonari.QPO — Type

QPO(S₀, f₀,A Q)

QPO model

S₀: the amplitude at the peak

f₀: the central frequency

Q: quality factor

\[\mathcal{P}(f) = \frac{S_0 {f_0}^4 } {\left(f^ 2 -{f_0}^2\right)^ 2 + f^2 {f_0}^2 / Q^2 }\]

Tonari.SingleBendingPowerLaw — Type

 SingleBendingPowerLaw(A, α₁, f₁, α₂)

Single bending power law model for the power spectral density

A: the amplitude

α₁: the first power law index

f₁: the first bend frequency

α₂: the second power law index

\[\mathcal{P}(f) = A \frac{(f/f₁)^{-α₁}}{1 + (f / f₁)^{α₂ - α₁}}\]

Tonari.separate_psd — Method

 separate_psd(psd::PowerSpectralDensity)

Separate the PSD into its BendingPowerLaw components and other components if it is a sum of PSDs

Arguments

psd::PowerSpectralDensity: power spectral density or sum of PowerSpectralDensity objects

Return

psd_continuum::Union{SumOfPowerSpectralDensity,PowerSpectralDensity,nothing}: continuum part of the psd

psd_line::Union{PowerSpectralDensity,nothing,Vector{PowerSpectralDensity}}: non-continuum part of the psd

Simulations

Tonari.Simulation — Type

Simulation(model::Model, T::Real, Δt::Real, S_high::Real, S_low::Real, t::AbstractVector{Real})

A struct that contains the information of a simulation of a stochastic process.

Fields

model::Model: The power spectral density of the process.

T::Real: The duration of the simulated time series. Note that the time stamps are from 0 to T, so the duration is T.

Δt::Real: The sampling period, or minimum time difference between samples.

S_high::Real: The factor by which the maximum frequency is multiplied for the simulation.

S_low::Real: The factor by which the minimum frequency is divided for the simulation.

t::AbstractVector{Real}: The time vector, in this case, the time at which the process is sampled. It is assumed that the time vector is sorted.

Constructors

Simulation(model::Model, T::Real, Δt::Real, S_high::Real, S_low::Real, t::AbstractVector{Real}): Constructs a simulation with regular sampling.

Simulation(model::Model, T::Real, Δt::Real): Constructs a simulation with regular sampling, and sets S_high and S_low to 10.0.

Simulation(model::Model, T::Real, Δt::Real, S_high::Real, S_low::Real): Constructs a simulation with the given parameters.

Simulation(model::Model, t::AbstractVector{Real}, S_high::Real, S_low::Real): Constructs a simulation with sampling pattern given by t.

Simulation(model::Model, t::AbstractVector{Real}): Constructs a simulation with sampling pattern given by t, and sets S_high and S_low to 10.0.

StatsBase.sample — Function

sample(rng, sim, n=1, input_mean=0; σₓ = nothing, split_long=false, randomise_values=true, Fvar=nothing, alt=false, poisson=false, exponentiate=false, error_size=0.02)

Generate a time series with a given power spectral density (PSD) using the (Timmer and König, 1995) method.

Arguments

rng::MersenneTwister: Random number generator.

sim::Simulation: The simulation

n::Int: Number of time series to generate. Default is 1.

input_mean::Real: The mean of the time series. Default is 0.

σₓ::Array{Float64, 1}: The errorbars of the time series. Default is nothing.

split_long::Bool: If true, the time series is split into shorter time series given by sim.S_low-1. Default is true.

randomise_values::Bool: If true, the values of the time series are randomised. Default is true.

Fvar::Real: The variance of the time series. Default is nothing.

alt::Bool: If true, uses the alternative Timmer & Koenig method. Default is false.

poisson::Bool: If true, Poisson noise is added to the time series. Default is false.

exponentiate::Bool: If true, the time series is exponentiated. Default is false.

error_size::Real: The size of the error. Default is 0.05.

Tonari.draw_errorbars — Function

draw_errorbars(rng, x, Δt, poisson=false, error_size=0.05)

Draw errorbars for the time series.

Arguments

rng::MersenneTwister: Random number generator.

x::Array{Float64, 1}: The values of the time series.

Δt::Real: The sampling period.

poisson::Bool: If true, Poisson noise is added to the time series. Default is false.

error_size::Real: The size of the error. Default is 0.05.

Tonari.findnearest — Method

findnearest(a, b)

Find the nearest value in b to each value in a. This assumes that a and b are sorted.

Tonari.randomise_fluxes — Function

randomise_fluxes(rng, xₛ, Δt, input_mean = 0.0; σ = nothing, Fvar = nothing, poisson = false, exponentiate = false, error_size = 0.05)

Randomise the fluxes of the time series.

Arguments

rng::MersenneTwister: Random number generator.

xₛ::Array{Float64, 1}: The values of the time series.

Δt::Real: The sampling period.

input_mean::Real: The mean of the time series. Default is 0.

σ::Array{Float64, 1}: The errorbars of the time series. Default is nothing.

Fvar::Real: The variance of the time series. Default is nothing.

poisson::Bool: If true, Poisson noise is added to the time series. Default is false.

exponentiate::Bool: If true, the time series is exponentiated. Default is false.

Tonari.sample_split_timeseries — Method

sample_split_timeseries(x, t, t_desired, n_sim, n, n_slices, split_long)

Split the time series into shorter time series and sample at the desired time stamps.

Arguments

x::Array{Float64, 1} or Vector{Matrix{Float64}}: The values of the time series. If there are multiple bands, the time series is a vector of matrices.

t::Array{Float64, 1}: The time indexes of the time series.

t_desired::Array{Float64, 1}: The desired time stamps.

n_sim::Int: The number of simulations.

n::Int: The number of time series to generate.

n_slices::Int: The number of slices to break the time series into.

split_long::Bool: If true, the time series is split into shorter time series. Default is true.

Returns

A tuple of two arrays: the first containing the time indexes of the shorter time series, and the second containing the values of the shorter time series.

Tonari.sample_timeseries — Method

sample_timeseries(t, y, M)

Extract a random subset of points from the time series.

Tonari.split_longtimeseries — Method

split_longtimeseries(t, ts, n_slices, t_end)

Split a long time series into shorter time series. Break the time series into n_slices shorter time series. The short time series are of equal length.

Arguments

t: The time indexes of the long time series.

ts: The values of the long time series.

n_slices: The number of slices to break the time series into.

t_end

Returns

A tuple of two lists: the first containing the time indexes of the shorter time series, and the second containing the values of the shorter time series.

Tonari.timmer_koenig — Method

timmer_koenig(psd, rng::random.AbstractRNG, alternative=false)

Generate a time series with a given power spectral density (PSD) using the (Timmer and König, 1995) method

Given N values of the power spectral density (PSD) 𝓟

Draw 2N values from a standard normal distribution.

The amplitude of the randomised periodogram is given by A = N + i M

The randomised periodogram is given by 𝓟_rand = √(𝓟 / 2) * A

The first value of the randomised periodogram inserted and set to 0

The time series is obtained by taking the inverse Fourier transform of the randomised periodogram, with the function FFTW.irfft. We use the length of the time series as 2*(N+1)-1.

The alternative parametrisation is given by

Given N values of the power spectral density (PSD) 𝓟

Draw N-1 values from a χ²(2) distribution and 1 value from χ²₁(1), this is A the amplitude of the randomised periodogram

Draw N values from a uniform distribution between 0 and 1, and set the last value to 0, this is θ the phase of the randomised periodogram

The randomised periodogram is given by 𝓟_rand = √(𝓟 / 2 * A) * exp(2πiθ)

The first value of the randomised periodogram inserted and set to 0

The time series is obtained by taking the inverse Fourier transform of the randomised periodogram, with the function FFTW.irfft. We use the length of the time series as 2*(N+1)-1.

Arguments

𝓟::Array{Float64, 1}: PSD associated with the process

rng::Random.AbstractRNG: Random number generator.

α::Real: Multiplicative factor for the randomised periodogram. Default is 1.0. This can be a complex vector to add a phase to the time series.

Returns

x::Array{Float64, 1}: Time series with the given PSD.

Periodogram

Tonari.coherence — Function

coherence(f, C̄, P̄₁, P̄₂, n_segments, y₁_err = nothing, y₂_err = nothing)

Compute the coherence between two time series defined in (Vaughan and Nowak, 1997) using the cross-periodogram C̄, the mean power spectrum of the first time series P̄₁, and the mean power spectrum of the second time series P̄₂. The coherence is defined as:

\[\gamma^2 (f) = \frac{|\overline{C}(f)|^2}{\overline{P}_1(f) \overline{P}_2(f)}\]

with the error in the coherence given by:

\[\sigma_{\gamma^2} = \sqrt{2} \sqrt{\gamma^2} \frac{1 - \gamma^2}{\sqrt{M}}\]

where M is the number of segments over which the coherence is computed.

The phase lag is defined as:

\[\Delta \phi (f) = \arg \overline{C}(f)\]

with the error in the phase lag given by:

\[\sigma_{\Delta \phi} = \frac{1 - \gamma^2}{\sqrt{2 \gamma^2 M}}\]

The time lag is defined as:

\[\Delta \tau (f) = \frac{\Delta \phi}{2 \pi f}\]

with the error in the time lag given by:

\[\sigma_{\Delta \tau} = \frac{\sigma_{\Delta \phi}}{2 \pi f}\]

If the errors in the time series are provided, the corrected coherence is computed using the method described in (Vaughan and Nowak, 1997).

Arguments

f::Array{Float64, 1}: Frequency array.

C̄::Array{Complex{Float64}, 1}: Cross-periodogram.

P̄₁::Array{Float64, 1}: Mean power spectrum of the first time series.

P̄₂::Array{Float64, 1}: Mean power spectrum of the second time series.

n_segments::Int: Number of segments over which the coherence is computed.

y₁_err::Array{Float64, 1}: Error in the first time series.

y₂_err::Array{Float64, 1}: Error in the second time series.

if the errors in the time series are not provided, the function returns the coherence, phase lag, and time lag. Otherwise, it returns the corrected coherence, phase lag, and time lag.

Returns

f::Array{Float64, 1}: Frequency array.

γ²::Array{Float64, 1}: Coherence.

Δφ::Array{Float64, 1}: Phase lag.

γ²_err::Array{Float64, 1}: Error in the coherence.

Δφ_err::Array{Float64, 1}: Error in the phase lag.

Δτ::Array{Float64, 1}: Time lag.

Δτ_err::Array{Float64, 1}: Error in the time lag.

P̄₁::Array{Float64, 1}: Mean power spectrum of the first time series.

Tonari.cross_periodogram — Function

cross_periodogram(t, y₁, y₂, y₁_err = nothing, y₂_err = nothing; compute_coherence = true, apply_end_matching = false, subtract_mean = true)

Compute the cross-periodogram between two time series y₁ and y₂ with time stamps t using the fast Fourier transform (FFT).

Arguments

t::Array{Float64, 1}: Time stamps.

y₁::Array{Float64, 1}: First time series data. Or a matrix of multiple time series.

y₂::Array{Float64, 1}: Second time series data. Or a matrix of multiple time series.

y₁_err::Array{Float64, 1}: Error in the first time series. Or a matrix of errors for multiple time series.

y₂_err::Array{Float64, 1}: Error in the second time series Or a matrix of errors for multiple time series.

compute_coherence::Bool: Compute the coherence between the two time series. Default is true.

apply_end_matching::Bool: Apply end-matching to the data. Default is false.

subtract_mean::Bool: Subtract the mean from the data. Default is true.

If compute_coherence is true, the function returns the coherence between the two time series. Otherwise, it returns the cross-periodogram and the mean power spectra of the two time series. see coherence function for the optional returns.

Returns

f::Array{Float64, 1}: Frequency array.

C̄::Array{Complex{Float64}, 1}: Mean cross-periodogram.

P̄₁::Array{Float64, 1}: Mean power spectrum of the first time series.

P̄₂::Array{Float64, 1}: Mean power spectrum of the second time series.

C::Array{Complex{Float64}, 1}: Cross-periodogram.

Tonari.end_matching — Method

end_matching(y,t)

End-match the data y with a straight line and subtract it from the data.

Tonari.periodogram — Function

periodogram(t,y, normalisation = "default"; apply_end_matching = false, subtract_mean = true)

Compute the periodogram of the data y with time stamps t using the fast Fourier transform (FFT).

The periodogram is computed as the squared magnitude of the Fourier transform of the data. In practrice we use the real-valued fast Fourier transform (rfft) to compute the periodogram.

Arguments

t::Array{Float64, 1}: Time stamps.

y::Array{Float64, 1}: Time series data.

normalisation::Float64: Normalisation factor. Default is 2Δt / length(t).

apply_end_matching::Bool: Apply end-matching to the data. Default is false.

subtract_mean::Bool: Subtract the mean from the data. Default is true.

Correlations

Tonari.cross_correlate — Method

cross_correlate(t₁::AbstractVector{T}, x₁::AbstractVector{T}, t₂::AbstractVector{T}, x₂::AbstractVector{T}; σ₁=nothing, σ₂=nothing, τ_list=nothing, Δτ=nothing, max_lag=nothing, local_estimate=false, both_ways=true, method="iccf", compute_errors=false, peak_frac=0.8, bootstrap=false, n_simulations=1_000,
skip_sanity_checks = false) where T

Compute the cross-correlation function between two time series.

The function computes the cross-correlation function between two time series by linearly interpolating the second time series using the interpolated cross-correlation function (ICCF) method. Most of this code is based on the R package sour by Simon Vaughan, University of Leicester https://github.com/SimonVaughanDataAndCode/sour.

Arguments

t₁::AbstractVector{T}: time stamps of the first time series

x₁::AbstractVector{T}: values of the first time series

t₂::AbstractVector{T}: time stamps of the second time series

x₂::AbstractVector{T}: values of the second time series

Optional arguments

σ₁::AbstractVector{T} = nothing: uncertainty of the first time series

σ₂::AbstractVector{T} = nothing: uncertainty of the second time series

τ_list::AbstractVector{T} = nothing: list of time lags, if not provided, it will be computed based on the time range of the time series

Δτ::T = nothing: time step between the time lags

max_lag::T = nothing: maximum time lag

local_estimate::Bool = false: flag to estimate the local mean and standard deviation of the time series

both_ways::Bool = true: Compute the cross-correlation function in both ways, i.e., from the first to the second time series and from the second to the first time series

method::String = "iccf": method to compute the cross-correlation function

compute_errors::Bool = false: flag to compute the errors of the cross-correlation function

peak_frac::T = 0.8: fraction of the peak to integrate to get the lag

bootstrap::Bool = false: flag to compute the errors using bootstrapping

n_simulations::Int = 1_000: number of simulations to compute the errors

skip_sanity_checks::Bool = false: flag to skip the sanity checks

Returns

τ_list::AbstractVector{T}: list of time lags

r::AbstractVector{T}: cross-correlation function

q::AbstractVector{T}: Distribution of centroid lags if compute_errors is true

Tonari.get_lag_list — Method

get_lag_list(Δτ, max_lag,t₁,t₂,τ_list)

Get the list of time lags to compute the cross-correlation function.

Arguments

Δτ::T: time step between the time lags

max_lag::T: maximum time lag

t₁::AbstractVector{T}: time stamps of the first time series

t₂::AbstractVector{T}: time stamps of the second time series

τ_list::AbstractVector{T}: list of time lags

Returns

τ_list::AbstractVector{T}: list of time lags

Δτ::T: time step between the time lags

max_lag::T: maximum time lag

n_lags::Int64: number of time lags

Tonari.iccf — Method

iccf(t₁::AbstractVector{T}, x₁::AbstractVector{T}, t₂::AbstractVector{T}, x₂::AbstractVector{T}, τ_list::AbstractVector{T} ; local_estimate=false) where T

Compute the interpolated cross-correlation function (ICCF) between two time series.

The function computes the cross-correlation function between two time series by linearly interpolating the second time series on the time points of the first time series. The ICCF is presented in Edelson et al. 2017 https://ui.adsabs.harvard.edu/abs/2017ApJ...840...41E/abstract and Peterson et al. 2004 https://ui.adsabs.harvard.edu/abs/2004ApJ...613..682P/abstract. The current implementation is based on the R package sour by Simon Vaughan, University of Leicester https://github.com/SimonVaughanDataAndCode/sour.

https://ui.adsabs.harvard.edu/abs/1986ApJ...305..175G/abstract

Arguments

t₁::AbstractVector{T}: time points of the first time series

x₁::AbstractVector{T}: values of the first time series

t₂::AbstractVector{T}: time points of the second time series

x₂::AbstractVector{T}: values of the second time series

τ_list::AbstractVector{T}: list of time lags

local_estimate::Bool=false: flag to estimate the local mean and standard deviation of the time series

Returns

CCF::Vector{T}: cross-correlation function

nx::Vector{Int64}: number of points in each time lag bin

Example

using Random
Random.seed!(123)
t₁ = 0:0.1:10
x₁ = randn(length(t₁))
t₂ = 0:0.1:10
x₂ = randn(length(t₂))
τ_list = -10:0.1:10
CCF,nx = iccf(t₁, x₁, t₂, x₂, τ_list)

Tonari.iccf_errors — Method

iccf_errors(rng::AbstractRNG, t₁::AbstractVector{T}, x₁::AbstractVector{T}, t₂::AbstractVector{T}, x₂::AbstractVector{T}, σ₁::AbstractVector{T} = nothing, σ₂::AbstractVector{T} = nothing,peak_frac = 0.8,local_estimate = false,τ_list = nothing,bootstrap = false,n_simulations = 1_000) where {T}

Compute the errors of the interpolated cross-correlation function (ICCF) between two time series.

Arguments

rng::AbstractRNG: Random number generator

t₁::AbstractVector{T}: time points of the first time series

x₁::AbstractVector{T}: values of the first time series

t₂::AbstractVector{T}: time points of the second time series

x₂::AbstractVector{T}: values of the second time series

σ₁::AbstractVector{T} = nothing: uncertainty of the first time series

σ₂::AbstractVector{T} = nothing: uncertainty of the second time series

peak_frac::T = 0.8: fraction of the peak to integrate to get the lag

local_estimate::Bool = false: flag to estimate the local mean and standard deviation of the time series

τ_list::AbstractVector{T} = nothing: list of time lags

bootstrap::Bool = false: flag to compute the errors using bootstrapping

n_simulations::Int = 1_000: number of simulations to compute the errors

Returns

q::Vector{T}: array of lags

Tonari.randomise_lc_flux — Method

randomise_lc_flux(rng::AbstractRNG, t::Vector{Float64}, x::Vector{Float64}; σ::Vector{Float64} = nothing, bootstrap::Bool = false)

Randomise the light curve by sampling with replacement. If bootstrap is true, the errors are not added to the light curve.

Arguments

rng::AbstractRNG: Random number generator

t::Vector{Float64}: Time array

x::Vector{Float64}: Flux array

σ::Vector{Float64}: Error array

bootstrap::Bool: If true, the errors are not added to the light curve

Returns

t_samp::Vector{Float64}: Sampled time array

x_samp::Vector{Float64}: Sampled flux array

σ_samp::Vector{Float64}: Sampled error array